Install python requirements
When building a Zephyr project, I realized that some Python modules were
missing. There are requirements*.txt files inside the Zephyr repo that
list these requirments, but since I didn't want the whole Zephyr source
code in the image, I had to find a way to get the requirements*.txt
files. That's what the multistage build is used for now.