Add date to image version tag
As it might happen that I build different versions of a container that
is based on the same SDK and Zephyr versions, I figured there has to be
a way to distinguish them. So I added the date at the end. I hope I
won't build multiple such containers on the same day.